requested changes

+++ src/app/shared/components/workshop-card/workshop-card.component.ts
@@ -1,6 +1,7 @@
 import { Component, EventEmitter, Input, OnDestroy, OnInit, Output } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
+import { ENTER, SPACE } from '@angular/cdk/keycodes';
 import { Select, Store } from '@ngxs/store';
 import { Observable, Subject } from 'rxjs';
 import { filter, takeUntil } from 'rxjs/operators';
@@ -111,7 +112,7 @@ export class WorkshopCardComponent implements OnInit, OnDestroy {
   }
 
   public onKeydown(event: KeyboardEvent, action: () => void): void {
-    if (event.key === 'Enter' || event.key === ' ') {
+    if (event.keyCode === ENTER || event.keyCode === SPACE) {
       action();
       event.preventDefault();
     }
@@ -166,14 +167,16 @@ export class WorkshopCardComponent implements OnInit, OnDestroy {
   }
 
   public onDisLike(): void {
-    this.store.dispatch([
-      new DeleteFavoriteWorkshop(this.favoriteWorkshopId),
-      new ShowMessageBar({
-        message: SnackbarText.deleteWorkshopFavorite,
-        type: 'success'
-      })
-    ]);
-    this.isFavorite = !this.isFavorite;
+    if (this.role === Role.parent) {
+      this.store.dispatch([
+        new DeleteFavoriteWorkshop(this.favoriteWorkshopId),
+        new ShowMessageBar({
+          message: SnackbarText.deleteWorkshopFavorite,
+          type: 'success'
+        })
+      ]);
+      this.isFavorite = !this.isFavorite;
+    }
   }
 
   public onChangeWorkshopStatus(status: string, type: ModalConfirmationType): void {
